pyproject.toml
[tool.poetry]
name = "cycls"
version = "0.0.2.55"

packages = [{ include = "cycls" }]
include = ["cycls/theme/**/*"]
description = "Cycls SDK"
authors = ["Mohammed J. AlRujayi <mj@cycls.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
httpx = "^0.27.0"
modal = "^1.1.0"
jwt = "^1.4.0"
docker = "^7.1.0"
cloudpickle = "^3.1.1"
openai = "^2.7.1"
google-generativeai = "^0.8.5"
pypdf = "^6.4.1"
python-docx = "^1.2.0"
pandas = "^2.3.3"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"